requirements.txt
# PDF Processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Vector Store and Embeddings
chromadb>=0.4.0
openai>=1.0.0

# Text Processing
tiktoken>=0.5.0
rank-bm25>=0.2.2

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Web Framework
flask>=3.0.0
